Move standalone some components in dedicated files
