+++ src/components/CategoryIcon.js
